// src/auth/decorators/access-strategy.enum.ts
export enum AccessStrategyEnum {
  ROLE_AND_ALL_PERMISSIONS = 'ROLE_AND_ALL_PERMISSIONS',
  ROLE_AND_ANY_PERMISSION = 'ROLE_AND_ANY_PERMISSION',
  ROLE_OR_ALL_PERMISSIONS = 'ROLE_OR_ALL_PERMISSIONS',
  ROLE_OR_ANY_PERMISSION = 'ROLE_OR_ANY_PERMISSION',
  ONLY_ROLE = 'ONLY_ROLE',
  ONLY_ALL_PERMISSIONS = 'ONLY_ALL_PERMISSIONS',
  ONLY_ANY_PERMISSION = 'ONLY_ANY_PERMISSION',
}